Use the windowed RANK() function in the voters example.
The query to calculate the heat rank in the voters example used to
calculate the heat map by rummaging through the rollup of votes by
contestant_number and state, keeping track of the winner and the total
number of votes.  This commit uses two SQL queries in one stored
procedure to avoid the java computation.  The first statement uses
RANK() to determine which contestant is the winner in a given state.
The second vote counts the number of votes.  The GetStateHeatmap stored
procedure returns both tables.

The dashboard index.html file has been updated to take advantage of
these changes.