Eng 10608 sqlcov join mview (#3889)
* Added new joined-matview test suite (joined-matview.sql,
joined-matview-DDL.sql, joined-matview-schema.py), and added it to
regression-config.py. To support that, split most of
advanced-template.sql off into a separate file, advanced-select.sql, so
that you can test its SELECT statements without its DELETE and INSERT
statements. And updated SQLGenerator.py, to allow specifying a 'count',
which is the number of random integers to generate (view names, in this
case), and a 'step', which lets you specify non-random integers, between
the specified min and max. (And trivial clean-up in basic-delete.sql,
basic-template.sql.)

* Moved joined-matview test suite to geo-config.py, and added new ones
(also in geo-config): joined-matview-int, joined-matview-string,
joined-matview-timestamp.

* Copied the existing joined-matview test suite into two versions:
joined-matview-default-quick and joined-matview-default-full, and
refactored to remove redundancy (joined-matview-default-template.sql,
basic-and-advanced-template.sql); kept the former in geo-config, but put
the latter in config.py (i.e., the 'extended' tests) - along with
joined-matview-int, joined-matview-string, joined-matview-timestamp; and
fixed joined-matview-timestamp-DDL.sql.